⚡️ Tries to add loading state to articles page
